✨ feat: Auto-fill wdio config path
Now, when you manually create a new run configuration, the "Wdio config file" field will be filled in automatically.

The plugin looks for common wdio config file names (like wdio.conf.ts or wdio.local.conf.js) in your project to find the right one.

To make this work, the logic for finding the config file was moved from WdioRunConfigurationProducer into a shared WdioUtil.findWdioConfig function.
